backend/pyproject.toml
[tool.poetry]
name = "hw-deal-sourcing-backend"
version = "1.0.0"
description = "ADK agents for HW Deal Sourcing - AI-driven investment opportunity discovery"
authors = ["HW <contact@hw-dealsourcing.com>"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
# Google Cloud and Agent Engine
google-cloud-aiplatform = {extras = ["adk", "agent-engines"], version = "^1.93.0"}
google-genai = "^1.9.0"
google-auth = "^2.0.0"
google-adk = "^1.0.0"
# Core dependencies
pydantic = "^2.10.6"
python-dotenv = "^1.0.1"
aiohttp = "^3.10.0"
requests = "^2.32.0"
# Document processing
reportlab = "^4.2.0"
markdown = "^3.7.0"
beautifulsoup4 = "^4.12.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
black = "^25.1.0"
pytest-asyncio = "^0.26.0"

[tool.poetry.group.deployment.dependencies]
absl-py = "^2.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
